\name{AlcanLink22}
\alias{AlcanLink22}
\docType{data}
\title{
ALCANLINK22 DATASET
}
\description{
  This dataset contains an extract from the linked ALCANLink data for calculating the
  incidence proportion.
}
\usage{data("AlcanLink22")}
\format{
  A data frame with 11463 observations on the following 13 variables.
  \describe{
    \item{\code{ID}}{a character vector}
    \item{\code{NEST_YR}}{a numeric vector}
    \item{\code{TOTCNT}}{a numeric vector}
    \item{\code{SAMCNT}}{a numeric vector}
    \item{\code{WTANAL}}{a numeric vector}
    \item{\code{STRATUMC}}{a numeric vector}
    \item{\code{SUD_NEST}}{a numeric vector}
    \item{\code{fpc}}{a numeric vector}
    \item{\code{FST_RPT_CensorAge}}{a numeric vector}
    \item{\code{FST_RPT_Indicator}}{a numeric vector}
    \item{\code{mchrace3}}{a character vector}
    \item{\code{public_health_region}}{a character vector}
    \item{\code{sex_code}}{a numeric vector}
  }
}
\source{
%%  ~~ R:\SCANlink\ALCANLink\Current linked files ~~
}

\keyword{datasets}
